work_02.02.2023-2

+++ src/main/java/by/iba/kazakov/device_verification/repositories/StandardRepository.java
@@ -3,5 +3,6 @@ package by.iba.kazakov.device_verification.repositories;
 import by.iba.kazakov.device_verification.models.Standard;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StandardRepository extends CrudRepository<Standard, Integer> {
+public interface StandardRepository extends CrudRepository<Standard, Long> {
+
 }
